Added exception package to model to better organize files.
