+++ app/src/main/java/classified/classifiedbuzzmovieselector/controllers/LoginActivity.java
@@ -10,8 +10,7 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import classified.classifiedbuzzmovieselector.R;
-import classified.classifiedbuzzmovieselector.model.InvalidPasswordException;
-import classified.classifiedbuzzmovieselector.model.UserAlreadyExistsException;
+import classified.classifiedbuzzmovieselector.model.Exceptions.UserAlreadyExistsException;
 import classified.classifiedbuzzmovieselector.model.UserManager;
 
 public class LoginActivity extends AppCompatActivity {
