+++ app/src/main/java/classified/classifiedbuzzmovieselector/model/Exceptions/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package classified.classifiedbuzzmovieselector.model;
+package classified.classifiedbuzzmovieselector.model.Exceptions;
 
 /**
  * Created by steven on 2/5/16.
